PR review improvements
