+++ client/state/selectors/is-blaze-enabled.js
@@ -1,10 +1,10 @@
 import { getSiteOption } from 'calypso/state/sites/selectors';
 
 /**
- * Returns if the currently selected site can Blaze posts.
+ * Returns if the specified site can Blaze posts.
  *
  * @param  {Object}  state Global state tree
- * @returns {?number}       Selected site ID
+ * @returns {boolean}  True if the site can blaze posts, false otherwise.
  */
 export default function isBlazeEnabled( state, siteId ) {
 	return !! getSiteOption( state, siteId, 'can_blaze' ) || false;
